+++ __tests__/app/models/index.js
@@ -1,13 +1,9 @@
-/* eslint-disable no-console */
-/* eslint-disable no-undef */
-/* eslint-disable no-alert */
 import {
 	Page,
 } from '../../../src/app';
 
 // eslint-disable-next-line import/extensions
-import User from './user.js';
-import UserTS from './user.ts';
+import User from './user';
 
 export default class Decorators extends Page {
 
@@ -15,14 +11,10 @@ export default class Decorators extends Page {
 		super.componentDidMount();
 		const user = new User();
 		try {
-			console.log(await user.get());
+			await user.load();
+			console.log(user);
 		} catch (e) {
-			alert(e.message || e);
-		}
-		const userTS = new UserTS();
-		try {
-			console.log(await userTS.get());
-		} catch (e) {
+			// eslint-disable-next-line no-undef
 			alert(e.message || e);
 		}
 	}